requirements.txt
streamlit
matplotlib
pandas
numpy
scikit-learn
opencv-python
Pillow
kaggle
